Feat: idea quick input with hashtag
